extract HardwareVideoEncoderWrapperFactory
