+++ sdk/android/BUILD.gn
@@ -531,6 +531,7 @@ if (is_android) {
       "api/org/webrtc/SimulcastVideoEncoderFactory.java",
       "api/org/webrtc/SimulcastAlignedVideoEncoderFactory.java",
       "api/org/webrtc/DefaultAlignedVideoEncoderFactory.java",
+      "api/org/webrtc/HardwareVideoEncoderWrapperFactory.java",
       "api/org/webrtc/HardwareVideoEncoderWrapper.java",
       "api/org/webrtc/ResolutionAdjustment.java",
     ]
