+++ sdk/android/api/org/webrtc/HardwareVideoEncoderWrapper.java
@@ -217,39 +217,4 @@ class HardwareVideoEncoderWrapper implements VideoEncoder {
     }
 }
 
-class HardwareVideoEncoderWrapperFactory implements VideoEncoderFactory {
-
-    private static final String TAG = "HardwareVideoEncoderWrapperFactory";
-
-    private final HardwareVideoEncoderFactory factory;
-    private final int resolutionPixelAlignment;
-
-    public HardwareVideoEncoderWrapperFactory(HardwareVideoEncoderFactory factory, int resolutionPixelAlignment) {
-        this.factory = factory;
-        this.resolutionPixelAlignment = resolutionPixelAlignment;
-        if (resolutionPixelAlignment == 0) {
-            throw new IllegalArgumentException("resolutionPixelAlignment should not be 0");
-        }
-    }
-
-    @Override
-    public VideoEncoder createEncoder(VideoCodecInfo videoCodecInfo) {
-        try {
-            VideoEncoder encoder = factory.createEncoder(videoCodecInfo);
-            if (encoder == null) {
-                return null;
-            }
-            return new HardwareVideoEncoderWrapper(encoder, resolutionPixelAlignment);
-        } catch (Exception e) {
-            Logging.e(TAG, "createEncoder failed", e);
-            return null;
-        }
-    }
-
-    @Override
-    public VideoCodecInfo[] getSupportedCodecs() {
-        return factory.getSupportedCodecs();
-    }
-}
-
